sourcemaps: fold the marker bridge into sourcemaps.js (drop src/markers.js)
markers.js had a single runtime consumer (compiler.js) and is thematically
source-map machinery — so fold MarkerRecorder + stripMarkers into sourcemaps.js
rather than carry a separate file or bloat the 5k-line compiler.js. compiler.js
now imports them from the sourcemaps module it already imports SourceMapGenerator
from; the marker sentinel consts are namespaced (MARK_OPEN/SHUT/SEP) to avoid
collision within the larger module.

No behavior change: markers byte-equivalence 18/18, corpus gap stays closed
(11/11), full test:all green, lints clean. Net: the RFC-12 fault-B work adds
zero new files (the marker primitive lives in the existing sourcemaps.js).
